linters: remove or move unused const/types/func
Many things are not needed when compiling for Windows, removing them or
moving them to the _linux only file.

Fixing:
	pkg/bpf/bpffs.go:105:2                                        unused       var mountOnce is unused
	pkg/cgroups/cgroups.go:36:6                                   unused       type deploymentEnv is unused
	pkg/observer/observer.go:34:2                                 unused       const perCPUBufferBytes is unused
	pkg/observer/observer.go:144:6                                unused       func perfBufferSize is unused
	pkg/observer/observer.go:159:20                               unused       func (*Observer).getRBSize is unused
	pkg/reader/namespace/namespace_windows.go:13:6                unused       type hostNamespaces is unused
	pkg/sensors/exec/procevents/proc_reader.go:26:2               unused       const maxMapRetries is unused
	pkg/sensors/exec/procevents/proc_reader.go:27:2               unused       const mapRetryDelay is unused
	pkg/sensors/exec/procevents/proc_reader_windows.go:183:6      unused       func updateExecveMapStats is unused
	pkg/sensors/program/loader_windows.go:22:2                    unused       var programTypeProcessGUID is unused
	pkg/sensors/sensor_types_windows.go:12:6                      unused       func sanitize is unused
	pkg/sensors/tracing/args.go:24:6                              unused       type argPrinter is unused
	pkg/sensors/tracing/args.go:33:2                              unused       const argReturnCopyBit is unused
	pkg/sensors/tracing/args.go:34:2                              unused       const argMaxDataBit is unused
	pkg/sensors/tracing/args.go:37:6                              unused       func argReturnCopy is unused
	pkg/sensors/tracing/args.go:47:6                              unused       func getMetaValue is unused
	pkg/sensors/tracing/args.go:67:6                              unused       func getTracepointMetaValue is unused
	pkg/sensors/tracing/args.go:77:6                              unused       func getArg is unused
	pkg/sensors/tracing/args.go:575:6                             unused       func parseString is unused
	pkg/sensors/tracing/consts.go:23:5                            unused       var errParseStringSize is unused
	pkg/sensors/tracing/consts.go:26:7                            unused       const maxStringSize is unused
	pkg/sensors/tracing/enforcer.go:50:6                          unused       func enforcerMapsUser is unused
	pkg/sensors/tracing/lists_windows.go:20:6                     unused       type listReader is unused
	pkg/sensors/tracing/lists_windows.go:24:23                    unused       func (*listReader).Read is unused
	pkg/sensors/tracing/loader_windows.go:5:6                     unused       func createLoaderEvents is unused
	pkg/sensors/tracing/policyhandler.go:23:6                     unused       type policyInfo is unused
	pkg/sensors/tracing/policyhandler.go:32:6                     unused       func newPolicyInfo is unused
	pkg/sensors/tracing/policyhandler.go:51:6                     unused       func newPolicyInfoFromSpec is unused
	pkg/sensors/tracing/policyhandler.go:71:23                    unused       func (*policyInfo).policyConfMap is unused
