+++ pkg/sensors/exec/procevents/proc_reader_windows.go
@@ -180,13 +180,6 @@ func getCWD(pid uint32) (string, uint32) {
 	return cwd, flags
 }
 
-func updateExecveMapStats(procs int64) {
-	//ToDo: WIP
-	// Currently we do not share the infor gathered in usermode with execve map in kernel in Windows,
-	// This method is currently stubbed out but will be implemented
-
-}
-
 func writeExecveMap(_ []procs) {
 	//ToDo: WIP
 	// Currently we do not share the infor gathered in usermode with execve map in kernel in Windows,
