+++ pkg/sensors/tracing/consts.go
@@ -3,8 +3,6 @@
 
 package tracing
 
-import "errors"
-
 const (
 	CharBufErrorENOMEM      = -1
 	CharBufErrorPageFault   = -2
@@ -19,20 +17,3 @@ const (
 	enforcerMapMaxEntries   = 32768
 	overrideMapMaxEntries   = 32768
 )
-
-var errParseStringSize = errors.New("error parsing string size from binary")
-
-// this is from bpf/process/types/basic.h 'MAX_STRING'
-const maxStringSize = 4096
-
-func kprobeCharBufErrorToString(e int32) string {
-	switch e {
-	case CharBufErrorENOMEM:
-		return "CharBufErrorENOMEM"
-	case CharBufErrorTooLarge:
-		return "CharBufErrorBufTooLarge"
-	case CharBufErrorPageFault:
-		return "CharBufErrorPageFault"
-	}
-	return "CharBufErrorUnknown"
-}
