--- pkg/sensors/tracing/loader_windows.go
@@ -1,8 +1,0 @@
-package tracing
-
-import "errors"
-
-func createLoaderEvents() error {
-
-	return errors.New("not supported on windows")
-}
